refactor: 프로필 이미지 관련 로직을 커스텀 훅으로 분리
- useProfileImage 커스텀 훅 생성
- 이미지 미리보기, 업로드, 삭제 로직을 훅으로 이동
- ProfileImageSection 컴포넌트 로직 단순화
